requirements.txt
streamlit==1.32.2
opencv-python==4.8.0.76
Pillow==10.2.0
numpy==1.26.4
